Move code block processing to new file
